Added Readme and better error handling to Autoinject Host Workshop (#22)
* initial commit of host based autoinjector

* Merge upstream/main and resolve conflicts

* Merge upstream/main and resolve conflicts

* Merge upstream/main and resolve conflicts

* add sample apps

* rebase

* rebase

* rebase

* feat(autoinject-host): Ensure OpenTelemetry config files are created

Previously, the OpenTelemetry Injector debian package did not consistently install the default configuration files (`java.conf`, `node.conf`, `dotnet.conf`) to `/etc/opentelemetry/otelinject`.

This commit modifies `01-install-autoinject-host.sh` to proactively check for the existence of these configuration files in the designated directory. If any are found to be missing, the script now creates them with their expected default content. This approach ensures that the auto-instrumentation can always be correctly configured after installation, irrespective of the deb package's behavior.

Additionally, this update includes a fix for an issue where the `dotnet.conf` file was not being correctly written due to a problem with multi-line string handling in the previous implementation.

* fixed the right flag

* removed the extra folder

* fix error handling

* fixed error handling

* feat(autoinject-host): Enhance installation scripts for OpenTelemetry Injector

This commit introduces several improvements to the installation scripts for the OpenTelemetry Injector. The `00-setup-env.sh` script now includes checks to verify if essential packages like `build-essential`, `ca-certificates`, and `curl` are installed before attempting to install them, preventing unnecessary installations.

The `01-install-autoinject-host.sh` script has been refactored to allow users to choose between a system-wide installation or a systemd-only installation method, enhancing flexibility. It also ensures that necessary configuration files are created if they do not already exist, and provides clear instructions for managing the installation.

Additionally, the `02-uninstall-autoinject-host.sh` script has been removed as part of a cleanup effort, streamlining the overall setup process.

These changes aim to improve user experience and ensure a more robust installation process for the OpenTelemetry Injector.

---------

Co-authored-by: root <root@ip-172-31-42-13.us-west-2.compute.internal>
Co-authored-by: Farhan Moshin <farhan.mohsin@coralogix.com>